+++ GBA/AdvancedBoy.hpp
@@ -20,7 +20,7 @@ bool InsertCartridge(fs::path romPath);
 
 /// @brief Run the emulator until the specified number of audio samples have been produced.
 /// @param samples Number of audio samples to produce.
-void FillAudioBuffer(int samples);
+void FillAudioBuffer(size_t samples);
 
 /// @brief Drain the emulator's internal audio buffer into the specified buffer.
 /// @param buffer Buffer to load internal audio buffer's samples into.
